pyproject.toml
[project]
name = "uniqimage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
]
